Make the signal handler thing work and add test case
