+++ api/bag.cpp
@@ -20,7 +20,6 @@
 #include "bag_vrrefinementsdescriptor.h"
 #include "bag_vrtrackinglist.h"
 #include <array>
-#include <csignal>
 #include <cstdio>
 #include <unistd.h>
 
@@ -36,135 +35,12 @@
 #include <sstream>
 #include <string>
 #include <string.h>
-#include <signal.h>
+#include "signal_hook.h"
 
 #ifdef _MSC_VER
 #pragma warning(pop)
 #endif
 
-
-namespace {
-
-static void (*cleanup_functions[128])(int);
-static int cleanup_functions_count = 0;
-
-// Returns 0 on success, 1 if too many functions have been set
-// This is not thread safe, only call it from a single thread
-int add_cleanup_function(void (*f)(int)) {
-    if (cleanup_functions_count >= 128) {
-        return 1;
-    }
-
-    cleanup_functions[cleanup_functions_count] = f;
-    cleanup_functions_count++;
-
-    return 0;
-}
-
-static void bag_handler(int signal);
-
-struct BagAbortHook {
-    //void (*old_handler[255])(int);
-    //void (*current_handler[255])(int);
-    struct sigaction old_handlers[255];
-    struct sigaction current_handlers[255];
-
-    // returns after done printing the exit message
-    void print_error_msg(int signal) {
-        // if the handlers are the same, don't call "bag handler" twice
-        char* signal_name;
-
-        #define TERM_STRING "a critical error occurred within BAG or a dependency (probably libxml or HDF5), exiting to prevent corruption or unexpected behavior\n"\
-                            "the signal that caused this error was: SIG"
-
-#ifdef __USE_GNU
-        const char* signal_abbrev = sigabbrev_np(signal);
-#else
-        const char* signal_abbrev = "<unknown>";
-#endif
-        if (!signal_abbrev) {
-            signal_abbrev = "<bad signal number>";
-        }
-
-        int stderr_fd = stderr->_fileno;
-
-        // strlen is not async signal safe, so it can't be used here
-        int signal_abbrev_length = 0;
-        for(;;signal_abbrev_length++) {
-            if (signal_abbrev[signal_abbrev_length] == 0) {
-                break;
-            }
-        }
-
-        //fputs(", FILE *__restrict stream);
-        write(stderr_fd, TERM_STRING, sizeof (TERM_STRING) - 1);
-        write(stderr_fd, signal_abbrev, signal_abbrev_length);
-    }
-
-    void call_cleanup_functions(int signal) {
-        for(int i = 0; i < cleanup_functions_count; i++) {
-            cleanup_functions[i](signal);
-        }
-    }
-
-    void call_previous_handler(int signal) {
-        if (this->old_handlers[signal].sa_handler != nullptr && this->old_handlers[signal].sa_handler != bag_handler) {
-            this->old_handlers[signal].sa_handler(signal);
-        }
-    }
-
-    void handle(int signal) {
-        this->print_error_msg(signal);
-
-        // if anyone set a cleanup function, call those
-        this->call_cleanup_functions(signal);
-
-        // if there was a handler before this, that isn't the BAG handler,
-        // call that one
-        this->call_previous_handler(signal);
-
-        // if nothing has exited in the previous handlers, exit here
-        _exit(-1);
-    }
-
-    std::array<int, 4> handled_signals() {
-        return { SIGABRT, SIGILL, SIGBUS, SIGFPE };
-    }
-
-    BagAbortHook() {
-        // set these all to null so that when restoring them
-        // we know which ones were added
-        for (int i = 0; i < 255; i++) {
-            this->old_handlers[i].sa_handler = nullptr;
-        }
-
-        struct sigaction old_action;
-        struct sigaction new_action;
-        for (auto sn : this->handled_signals()) {
-            new_action.sa_handler = bag_handler;
-            sigaction(sn, &new_action, &old_action);
-
-            this->old_handlers[sn] = old_action;
-        }
-
-    }
-
-    ~BagAbortHook() {
-        // reset the handlers back to what they were
-        for (auto sn : this->handled_signals()) {
-            struct sigaction old_action = this->old_handlers[sn];
-            sigaction(sn, &old_action, nullptr);
-        }
-    }
-};
-
-static BagAbortHook* current_hook;
-
-static void bag_handler(int signal) {
-    if (current_hook) {
-        current_hook->handle(signal);
-    }
-}
 
 //! Convert a BAG::CompoundDataType (C++) into a BagCompoundDataType (C).
 /*!
@@ -236,8 +112,6 @@ BAG::CompoundDataType getValue(
         return {};
     }
 }
-
-}  // namespace
 
 //! Open the specified BAG.
 /*!
