+++ examples/driver.cpp
@@ -1,7 +1,8 @@
+
+#include <string>
 #include <cstring>
 #include <iostream>
 #include <fstream>
-#include <string>
 #include <ios>
 #include <vector>
 #include <stdint.h>
